rag-frontend/requirements.txt
pydantic==2.10.1
chainlit==1.3.2
python-dotenv==1.0.1